nua-build/pyproject.toml
[tool.poetry]
name = "nua-build"
version = "0.4.33"
description = "Nua build package (currently: build, core build, runtime)"
authors = ["Stefane Fermigier <sf@abilian.com>", "Jerome Dumonteil <jd@abilian.com>"]
license = "AGPL"
readme = "README.md"
packages = [
  { include = "nua", from = "src" }
]

[tool.poetry.scripts]
nua-build = "nua.build.main:app"
nua_build_self_setup = "nua.build.scripts.nuad_self_setup:main"
nua_build_setup_app = "nua.build.scripts.nua_build_setup_app:main"
nua_test_replace_domain = "nua.build.scripts.test_replace_domain:main"

[tool.poetry.dependencies]
# nua---lib = { path = "../nua-lib/", develop = false }
nua-lib = "==0.4.33"
nua-runtime = "==0.4.33"
python = "^3.10"
setuptools = "^65.5.0"
tomli = "^2.0"
typer = {extras = ["all"], version = "^0.7.0"}
rich = "^12.4.4"
docker = "^6.0"
psycopg2-binary = "^2.9.4"  # fixme: need until refactoring
# mariadb  # same

[tool.poetry.group.dev.dependencies]

# Testing
pytest = "^7"
pytest-cov = "^4"
pytest-xdist = "*"
nox = "^2022.8.7"

# Linting
flake8 = "^5"
flake8-bandit = "*"
flake8-bugbear = "*"
flake8-cognitive-complexity = "*"
flake8-comprehensions = "*"
flake8-datetimez = "*"
flake8-functions = "*"
flake8-mutable = "*"
flake8-pytest = "*"
flake8-simplify = "*"
flake8-super = "*"
flake8-super-call = "*"
flake8-tidy-imports = "*"
flake8-tuple = "*"
pep8-naming = "*"
# Buggy plugins or not relevant for now
# bandit = "*"
# mastool = "^0.1.3"
# mccabe = "^0.7"
# flake8-aaa = "*"

# Formatting
isort = "*"
black = "*"
docformatter = "^1.4"

# Typechecking
mypy = "*"
types-toml = "*"

# More checks
vulture = "*"
deptry = "*"
ruff = "*"
pyright = "^1.1.279"


[tool.pytest.ini_options]
filterwarnings = ["ignore:.*distutils package:DeprecationWarning", "ignore::DeprecationWarning"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
ignore = [
    'E711', # comparison to None should be 'if cond is not None:'
    # TODO: fix these
    'E501', # Line too long
    'F401', # "..." imported but unused
    'F841', # Local variable assigned to but never used
    # Not ruff
    # 'W503', # line break before binary operator
    # 'S101', # Use of assert detected
]
exclude = [
    "tests/apps",
]


[tool.deptry]
exclude = [
    '.tox',
    '.nox',
    'tests',
    'sandbox',
    'doc',
    'scripts',
    "noxfile.py"
]
ignore_obsolete = [
    "setuptools",
]
ignore_missing = [
    "mariadb",
    "psycopg2",
]
ignore_transitive = [
]


[tool.pyright]
exclude = [
    '.tox',
    '.nox',
    'tests',
    'sandbox',
    'doc',
    'scripts',
    # Temp: dependencies are not loaded
    "nua_build/runtime/mariadb_utils.py",
    "nua_build/runtime/postgres.py",
]
